Integrate JWT identity, pricing, and DB integrity
Major refactor to enforce owner identity, decimal pricing and DB integrity across the orders service. Adds request/response DTOs with validation, replaces int prices with BigDecimal, and persists owner_issuer/owner_subject and optimistic version in Order; includes Liquibase changelog to migrate types and add new columns/index. CatalogClient now forwards bearer tokens, validates responses and uses timeouts. SecurityConfig and controllers refactored to use JwtAuthenticationToken and centralized authorization; OrderService rewritten to price items from catalog, enforce state transitions, ownership visibility and produce ProblemDetail/ResponseStatusException errors. Also adds global exception handler, unit/integration tests, README, Docker/compose and application property defaults, and small pom dependency fixes.

+++ src/main/java/com/cloudnative/ms_orders/dto/OrderItemResponseDTO.java
@@ -1,8 +1,3 @@
 package com.cloudnative.ms_orders.dto;
-
-public record OrderItemResponseDTO(
-    int productId,
-    int quantity,
-    int unitPrice,
-    ProductDTO product
-) {}
+import java.math.BigDecimal;
+public record OrderItemResponseDTO(int productId, int quantity, BigDecimal unitPrice, ProductDTO product) {}
