+++ src/main/java/com/cloudnative/ms_orders/dto/OrderResponseDTO.java
@@ -1,16 +1,8 @@
 package com.cloudnative.ms_orders.dto;
-
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
-
 import com.cloudnative.ms_orders.model.OrderStatus;
-
-public record OrderResponseDTO(
-    int id,
-    int customerId,
-    OrderStatus status,
-    List<OrderItemResponseDTO> items,
-    int totalAmount,
-    LocalDateTime createdAt,
-    LocalDateTime updatedAt
-) {}
+public record OrderResponseDTO(Long id, int customerId, OrderStatus status,
+    List<OrderItemResponseDTO> items, BigDecimal totalAmount,
+    LocalDateTime createdAt, LocalDateTime updatedAt) {}
